\name{PowerUpR-deprecated}
\alias{PowerUpR-deprecated}
\title{Deprecated and Defunct functions in \sQuote{PowerUpR}}

\description{
  Deprecated functions are provided for compatibility purposes and will be defunct in the next \sQuote{PowerUpR} release. Defunct functions are no longer supported.
}

\details{

  Deprecated and defunct functions and their replacement are as follows:

    \itemize{

      \item{\code{mrss.ira1r1} is deprecated, use \code{\link{optimal.ira1r1}}}
      \item{\code{mrss.cra2r2} is deprecated, use \code{\link{optimal.cra2r2}}}
      \item{\code{mrss.cra3r3} is deprecated, use \code{\link{optimal.cra3r3}}}
      \item{\code{mrss.cra4r4} is deprecated, use \code{\link{optimal.cra4r4}}}
      \item{\code{mrss.bcra3f2} is deprecated, use \code{\link{optimal.bcra3f2}}}
      \item{\code{mrss.bcra3r2} is deprecated, use \code{\link{optimal.bcra3r2}}}
      \item{\code{mrss.bcra4r2} is deprecated, use \code{\link{optimal.bcra4r2}}}
      \item{\code{mrss.bcra4f3} is deprecated, use \code{\link{optimal.bcra4f3}}}
      \item{\code{mrss.bcra4r2} is deprecated, use \code{\link{optimal.bcra4r2}}}
      \item{\code{mrss.bira4r1} is deprecated, use \code{\link{optimal.bira4r1}}}
      \item{\code{mrss.bira3r1} is deprecated, use \code{\link{optimal.bira3r1}}}
      \item{\code{mrss.bira2c1} is deprecated, use \code{\link{optimal.bira2c1}}}
      \item{\code{mrss.bira2f1} is deprecated, use \code{\link{optimal.bira2f1}}}
      \item{\code{mrss.bira2r1} is deprecated, use \code{\link{optimal.bira2r1}}}
      \item{\code{mrss.to.mdes} is defunct, use \code{\link{optimal.to.mdes}}}
      \item{\code{mrss.to.power} is defunct, use \code{\link{optimal.to.power}}}
      \item{\code{t1t2.error} is defunct, there is no replacement function}
      \item{\code{plot.pars} is defunct, there is no replacement function}

    }
}

\format{Deprecated functions are no longer documented, please use documentation for new functions.}

\keyword{deprecated or defunct}
