\docType{package}
\name{PowerUpR-package}
\alias{PowerUpR}
\alias{PowerUpR-package}
\title{Power Analysis Tools for Multilevel Randomized Experiments}
\description{
  \emph{PowerUp!} series consist of three excel-based applications to design various multilevel randomized experiments to detect main treatment effects, and to design two- and three-level cluster-randomized trials (CRTs) to detect multilevel moderation and mediation. For more information please refer to \url{https://www.causalevaluation.org/}.
}
