\name{PowerUpR-deprecated}
\title{Deprecated and Defunct functions in \pkg{PowerUpR}}
\alias{PowerUpR-deprecated}
\alias{mdes.med211}
\alias{mdes.med221}

\description{
 Experimental MDES functions for 2-1-1 and 2-2-1 mediations are removed.
}

\details{

  Defunct functions:

    \itemize{
      \item{\code{mdes.med211} is defunct, there is no replacement function}
      \item{\code{mdes.med221} is defunct, there is no replacement function}
    }
}

\format{Deprecated or defunct functions are no longer documented.}

