\name{PowerUpR-deprecated}
\title{Deprecated and Defunct functions in \pkg{PowerUpR}}
\alias{PowerUpR-deprecated}
\alias{optimal.ira1r1}
\alias{optimal.bira2c1}
\alias{optimal.bira2f1}
\alias{optimal.bira2r1}
\alias{optimal.cra2r2}
\alias{optimal.bira3r1}
\alias{optimal.bcra3f2}
\alias{optimal.bcra3r2}
\alias{optimal.cra3r3}
\alias{optimal.bira4r1}
\alias{optimal.bcra4r2}
\alias{optimal.bcra4f3}
\alias{optimal.bcra4r3}
\alias{optimal.cra4r4}
\alias{optimal.to.mdes}
\alias{optimal.to.power}


\description{
 Constrained optimal sample allocation functions are removed from \pkg{PowerUpR} package, see \pkg{cosa} package.
}

\details{

  Deprecated and defunct functions and their replacement are as follows:

    \itemize{

      \item{\code{optimal.ira1r1} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.cra2r2} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.cra3r3} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.cra4r4} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.bcra3f2} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.bcra3r2} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.bcra4r2} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.bcra4f3} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.bcra4r2} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.bira4r1} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.bira3r1} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.bira2c1} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.bira2f1} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.bira2r1} is defunct, see \pkg{cosa} package}
      \item{\code{optimal.to.mdes} is defunct,  there is no replacement function}
      \item{\code{optimal.to.power} is defunct, there is no replacement function}

    }
}

\format{Deprecated or defunct functions are no longer documented.}

