\docType{package}
\name{PowerUpR-package}
\alias{PowerUpR}
\alias{PowerUpR-package}
\title{Power Analysis Tools for Multilevel Randomized Experiments}
\description{
  \emph{PowerUp!} series consist of three excel-based applications which allow researchers to conduct statistical power analysis for various multilevel randomized experiments to detect main treatment effects, for two- and three-level cluster-randomized trials (CRTs) to detect moderation effects and for two-level CRTs to detect 2-2-1 and 2-1-1 mediation effects. For more information please refer to \url{http://www.causalevaluation.org/}.
}
