\docType{package}
\name{PowerUpR-package}
\alias{PowerUpR}
\alias{PowerUpR-package}
\title{Power Analysis Tools for Multilevel Randomized Experiments}
\description{
  \emph{PowerUp!} series consists of three excel-based applications that allows researchers to conduct statistical power analysis for main, moderator and mediator effects for various multilevel randomized experiments. For more information please refer to \url{http://www.causalevaluation.org/}.
}
