\docType{package}
\name{PowerUpR-package}
\alias{PowerUpR}
\alias{PowerUpR-package}
\title{PowerUpR: R version of \emph{PowerUp!}}
\description{
  \emph{PowerUp!} series consists of three excel-based applications that allows researchers to conduct statistical power analysis for main, moderator and indirect effects for various multilevel randomized experiments. For more information please refer to \url{http://www.causalevaluation.org/}.
}

\author{
  \strong{Authors}: Metin Bulus and Nianbo Dong
}
