\docType{package}
\name{PowerUpR-package}
\alias{PowerUpR}
\alias{PowerUpR-package}
\title{PowerUpR: R version of \emph{PowerUp!}}
\description{
  \emph{PowerUp!} series consists of three excel-based applications that allows researchers to conduct statistical power analysis for main effects (download link: \href{http://www.causalevaluation.org/uploads/7/3/3/6/73366257/powerup.xlsm}{\emph{PowerUp!}}), moderator effects (download link: \href{http://www.causalevaluation.org/uploads/7/3/3/6/73366257/powerup-moderator.xlsm}{\emph{PowerUp!-Moderator}}) and mediator effects (download link: \href{http://www.causalevaluation.org/uploads/7/3/3/6/73366257/powerup-mediator.xlsm}{\emph{PowerUp!-Mediator}}) for various multilevel randomized experiments and quasi-experiments. For more information please refer to \url{http://www.causalevaluation.org/}.
}

\author{
  \strong{Maintainer}: Metin Bulus \email{bulus.metin@gmail.com}

  Authors:
    \itemize{
      \item Nianbo Dong \email{dong.nianbo@rgmail.com}
      \item Benjamin Kelcey \email{ben.kelcey@gmail.com}
      \item Jessaca K. Spybrook \email{jessaca.spybrook@wmich.edu}
      \item Bang T. Lee \email{bangthanh.le@wmich.edu}
    }
}
\keyword{PowerUpR}
